add safe actions to dev page and remove safe settings page
